Refactor early lowering, add division
TODO: fix costs.

Coming this far, I am starting to believe
this not worth it. The diffs are practically
non-existent, while adding the plumbing adds
complexity and slows down the compiler.

The costs are not the only area where new code
is required. CSE also looks at calls for the
"live across call" bit, loop alignment heuristic
does so too. And those cause real, though minor,
code size regressions.

Ultimately, the only phases that are really helped
by this change are morph and assertion propagation.
There are also potential gains for fast division.

In morph, we'll be able to do more post-order
transformations with the operators. Assertion
propagation will be able to fold them when their
operands are determined to be constant. We'll also
be more flexible in allowing some division though,
as today, for example, even simple UMODs by power
of two get turned into helpers sometimes.

The drawbacks, however, seem to outweight the benefits:

1. New phase, which means TP impact. Rough and early PIN
   measurements suggest that the impact is around the ~0.2%
   mark, which is extremely disproportinal to the CQ benefits,
   and not in a good way. The new phase will subsume "simple
   lowering", so some of this will be recovered, but I do not
   anticipate getting it lower than 0.1%, perhaps 0.05%, and
   that is simply not acceptable.
2. Code for "am I equivalent to the call" goes against the spirit
   of the change, yet is required in at least 3 phases. Put simply,
   why are we trying to move the morphing later if early morphing
   is actually what we want?

The first point is a practical nail, the second one is conceptual,
but both are for the same coffin.

A bit sad, I expected to get a lot of benefits from this phase,
like, for example, late morphing for intrinsics so that constant
operands for them can be discovered in the optimizer, but seems
like that's not possible at this point, so we'll have to wait
until "fgMorphArgs" becomes available in lowering proper.

All paths lead to "fgMorphArgs", don't they :).