Centralize operand traversal
GenTreeVisitor and VisitOperands had each a copy of the
traversal code. Unify them through a new VisitUses method.

This change has been validated to not affect the performance
of either VisitOperands or GenTreeVisitor via microbenchmarks.

This required the use of FORCEINLINE (in the past, the code
was effectively copied and inlined by hands). It is likely we
can drop the FORCEINLINE and rely on PGO data for aggressive
inlining in this case, but that is yet to be validated, and
without the inlining there are regressions (on the scale of
20-50% slower, depending on the visitor).